Adds tons of changes: remove unused fields in psychology and physio exercises, add views and crud operations on foods and exercises, some PDF reports changes, fix minor errors, remove food_fields in Meal form and UI fixes

+++ db/migrate/20200107182936_create_foods.rb
@@ -8,7 +8,7 @@ class CreateFoods < ActiveRecord::Migration[6.0]
       t.float :protein
       t.float :carbs
       t.float :fats
-      t.float :serving_size
+      t.integer :serving_size
       t.boolean :vegan, default: false, null: false
       t.boolean :vegetarian, default: false, null: false
       t.boolean :gluten_free, default: false, null: false
